requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
asttokens==2.2.1
    # via stack-data
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via pytest
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.2
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
build==0.10.0
    # via pip-tools
celery==5.2.7
    # via
    #   -r requirements.in
    #   flower
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
decorator==5.1.1
    # via ipython
django==4.1.7
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-model-utils
    #   django-stubs
    #   django-stubs-ext
django-environ==0.10.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-model-utils==4.3.1
    # via -r requirements.in
django-stubs==1.15.0
    # via -r requirements.in
django-stubs-ext==0.7.0
    # via django-stubs
exceptiongroup==1.1.1
    # via pytest
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via -r requirements.in
faker==17.6.0
    # via factory-boy
flower==1.2.0
    # via -r requirements.in
humanize==4.6.0
    # via flower
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
ipython==8.11.0
    # via -r requirements.in
jedi==0.18.2
    # via ipython
kombu==5.2.4
    # via celery
lxml==4.9.2
    # via -r requirements.in
matplotlib-inline==0.1.6
    # via ipython
mypy==1.1.1
    # via
    #   -r requirements.in
    #   django-stubs
mypy-extensions==1.0.0
    # via mypy
packaging==23.0
    # via
    #   build
    #   pytest
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.12.3
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
prometheus-client==0.16.0
    # via flower
prompt-toolkit==3.0.38
    # via
    #   click-repl
    #   ipython
psycopg2==2.9.5
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.14.0
    # via ipython
pyproject-hooks==1.0.0
    # via build
pytest==7.2.2
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   faker
pytz==2022.7.1
    # via
    #   celery
    #   flower
redis==4.5.1
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
sentry-sdk==1.16.0
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
soupsieve==2.4
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
stack-data==0.6.2
    # via ipython
tomli==2.0.1
    # via
    #   build
    #   django-stubs
    #   mypy
    #   pytest
tornado==6.2
    # via flower
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
types-beautifulsoup4==4.11.6.7
    # via -r requirements.in
types-html5lib==1.1.11.12
    # via types-beautifulsoup4
types-python-dateutil==2.8.19.10
    # via -r requirements.in
types-pytz==2022.7.1.2
    # via django-stubs
types-pyyaml==6.0.12.8
    # via django-stubs
typing-extensions==4.5.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
urllib3==1.26.15
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
